✨ listen to table changes with live query
