+++ SurrealDb.Net.Tests/LiveQueryTests.Properties.cs
@@ -1,0 +1,6 @@
+﻿namespace SurrealDb.Net.Tests;
+
+public partial class LiveQueryTests
+{
+    private static readonly TimeSpan _timeout = TimeSpan.FromSeconds(2);
+}
