+++ SurrealDb.Net.Tests/LiveTests.cs
@@ -1,0 +1,181 @@
+﻿using SurrealDb.Net.Models.LiveQuery;
+using JsonPatchOperations = System.Collections.Immutable.IImmutableList<Microsoft.AspNetCore.JsonPatch.Operations.Operation>;
+
+namespace SurrealDb.Net.Tests;
+
+public class LiveTests
+{
+    [Fact]
+    public async Task ShouldNotBeSupportedOnHttpProtocol()
+    {
+        const string url = "http://localhost:8000";
+
+        Func<Task> func = async () =>
+        {
+            await using var surrealDbClientGenerator = new SurrealDbClientGenerator();
+            var dbInfo = surrealDbClientGenerator.GenerateDatabaseInfo();
+
+            using var client = surrealDbClientGenerator.Create(url);
+            await client.SignIn(new RootAuth { Username = "root", Password = "root" });
+            await client.Use(dbInfo.Namespace, dbInfo.Database);
+
+            await using var liveQuery = await client.Live<int>("test");
+        };
+
+        await func.Should().ThrowAsync<NotSupportedException>();
+    }
+
+    [Fact]
+    public async Task ShouldReceiveData()
+    {
+        const string url = "ws://localhost:8000/rpc";
+
+        var allResults = new List<SurrealDbLiveQueryResponse>();
+
+        Func<Task> func = async () =>
+        {
+            await using var surrealDbClientGenerator = new SurrealDbClientGenerator();
+            var dbInfo = surrealDbClientGenerator.GenerateDatabaseInfo();
+
+            using var client = surrealDbClientGenerator.Create(url);
+            await client.SignIn(new RootAuth { Username = "root", Password = "root" });
+            await client.Use(dbInfo.Namespace, dbInfo.Database);
+
+            var liveQuery = await client.Live<TestRecord>("test");
+
+            var cts = new CancellationTokenSource();
+
+            _ = Task.Run(async () =>
+            {
+                await foreach (var result in liveQuery.WithCancellation(cts.Token))
+                {
+                    allResults.Add(result);
+                }
+            });
+
+            _ = Task.Run(async () =>
+            {
+                await Task.Delay(50);
+
+                var record = await client.Create("test", new TestRecord { Value = 1 });
+                await client.Upsert(new TestRecord { Id = record.Id, Value = 2 });
+                await client.Delete(record.Id!);
+
+                while (allResults.Count < 3)
+                {
+                    await Task.Delay(20);
+                }
+
+                await liveQuery.KillAsync();
+
+                while (allResults.Count < 4)
+                {
+                    await Task.Delay(20);
+                }
+
+                cts.Cancel();
+            });
+
+            await Task.Delay(TimeSpan.FromSeconds(2));
+
+            if (!cts.IsCancellationRequested)
+            {
+                cts.Cancel();
+                throw new Exception("Timeout");
+            }
+        };
+
+        await func.Should().NotThrowAsync();
+
+        allResults.Should().HaveCount(4);
+
+        var firstResult = allResults[0];
+        firstResult.Should().BeOfType<SurrealDbLiveQueryCreateResponse<TestRecord>>();
+
+        var secondResult = allResults[1];
+        secondResult.Should().BeOfType<SurrealDbLiveQueryUpdateResponse<TestRecord>>();
+
+        var thirdResult = allResults[2];
+        thirdResult.Should().BeOfType<SurrealDbLiveQueryDeleteResponse>();
+
+        var lastResult = allResults[3];
+        lastResult.Should().BeOfType<SurrealDbLiveQueryCloseResponse>();
+    }
+
+    [Fact]
+    public async Task ShouldReceiveDataInJsonPatchFormat()
+    {
+        const string url = "ws://localhost:8000/rpc";
+
+        var allResults = new List<SurrealDbLiveQueryResponse>();
+
+        Func<Task> func = async () =>
+        {
+            await using var surrealDbClientGenerator = new SurrealDbClientGenerator();
+            var dbInfo = surrealDbClientGenerator.GenerateDatabaseInfo();
+
+            using var client = surrealDbClientGenerator.Create(url);
+            await client.SignIn(new RootAuth { Username = "root", Password = "root" });
+            await client.Use(dbInfo.Namespace, dbInfo.Database);
+
+            var liveQuery = await client.Live<JsonPatchOperations>("test", true);
+
+            var cts = new CancellationTokenSource();
+
+            _ = Task.Run(async () =>
+            {
+                await foreach (var result in liveQuery.WithCancellation(cts.Token))
+                {
+                    allResults.Add(result);
+                }
+            });
+
+            _ = Task.Run(async () =>
+            {
+                await Task.Delay(50);
+
+                var record = await client.Create("test", new TestRecord { Value = 1 });
+                await client.Upsert(new TestRecord { Id = record.Id, Value = 2 });
+                await client.Delete(record.Id!);
+
+                while (allResults.Count < 3)
+                {
+                    await Task.Delay(20);
+                }
+
+                await liveQuery.KillAsync();
+
+                while (allResults.Count < 4)
+                {
+                    await Task.Delay(20);
+                }
+
+                cts.Cancel();
+            });
+
+            await Task.Delay(TimeSpan.FromSeconds(2));
+
+            if (!cts.IsCancellationRequested)
+            {
+                cts.Cancel();
+                throw new Exception("Timeout");
+            }
+        };
+
+        await func.Should().NotThrowAsync();
+
+        allResults.Should().HaveCount(4);
+
+        var firstResult = allResults[0];
+        firstResult.Should().BeOfType<SurrealDbLiveQueryCreateResponse<JsonPatchOperations>>();
+
+        var secondResult = allResults[1];
+        secondResult.Should().BeOfType<SurrealDbLiveQueryUpdateResponse<JsonPatchOperations>>();
+
+        var thirdResult = allResults[2];
+        thirdResult.Should().BeOfType<SurrealDbLiveQueryDeleteResponse>();
+
+        var lastResult = allResults[3];
+        lastResult.Should().BeOfType<SurrealDbLiveQueryCloseResponse>();
+    }
+}
