+++ SurrealDb.Net.Tests/SurrealDb.Net.Tests.csproj
@@ -11,6 +11,7 @@
     <ItemGroup>
         <PackageReference Include="FluentAssertions" Version="6.12.0" />
         <PackageReference Include="Meziantou.Xunit.ParallelTestFramework" Version="2.1.0" />
+        <PackageReference Include="Microsoft.AspNetCore.JsonPatch" Version="7.0.13" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.7.2" />
         <PackageReference Include="xunit" Version="2.5.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.5.1">
